adding FD test case
