+++ applications/SystemIdentificationApplication/tests/responses/auxiliary_files/StructuralMaterials.json
@@ -1,0 +1,18 @@
+{
+    "properties" : [{
+        "model_part_name" : "Structure.Parts_Solid_full",
+        "properties_id"   : 1,
+        "Material"        : {
+            "constitutive_law" : {
+                "name" : "LinearElasticPlaneStrain2DLaw"
+            },
+            "Variables"        : {
+                "THICKNESS"     : 0.02,
+                "DENSITY"       : 7850.0,
+                "YOUNG_MODULUS" : 20.0,
+                "POISSON_RATIO" : 0.29
+            },
+            "Tables"           : null
+        }
+    }]
+}
