+++ applications/SystemIdentificationApplication/tests/responses/auxiliary_files/adjoint_material_parameters.json
@@ -1,0 +1,3 @@
+{
+    "properties": []
+}
