+++ applications/SystemIdentificationApplication/tests/responses/auxiliary_files/adjoint_project_parameters.json
@@ -1,0 +1,76 @@
+{
+    "problem_data": {
+        "problem_name": "benchmark_2",
+        "parallel_type": "OpenMP",
+        "start_time": 0,
+        "end_time": 1,
+        "echo_level": 0
+    },
+    "sensor_settings": {
+        "perturbation_size": 1e-8,
+        "adapt_perturbation_size": true,
+        "@include_json": "auxiliary_files/sensor_data.json"
+    },
+    "solver_settings": {
+        "solver_type": "adjoint_static",
+        "analysis_type": "linear",
+        "model_part_name": "AdjointStructure",
+        "domain_size": 3,
+        "time_stepping": {
+            "time_step": 1.0
+        },
+        "compute_reactions": false,
+        "move_mesh_flag": false,
+        "sensitivity_settings": {
+            "sensitivity_model_part_name": "Parts_Solid_full",
+            "element_data_value_sensitivity_variables": [
+                "YOUNG_MODULUS"
+            ],
+            "build_mode": "static"
+        },
+        "echo_level": 0,
+        "rotation_dofs": true,
+        "model_import_settings": {
+            "input_type": "use_input_model_part"
+        },
+        "material_import_settings": {
+            "materials_filename": "auxiliary_files/adjoint_material_parameters.json"
+        },
+        "linear_solver_settings": {
+            "solver_type": "amgcl"
+        }
+    },
+    "processes": {
+        "constraints_process_list": [
+            {
+                "python_module": "assign_vector_variable_process",
+                "kratos_module": "KratosMultiphysics",
+                "help": "This process fixes the selected components of a given vector variable",
+                "process_name": "AssignVectorVariableProcess",
+                "Parameters": {
+                    "mesh_id": 0,
+                    "model_part_name": "AdjointStructure.DISPLACEMENT_left",
+                    "variable_name": "ADJOINT_DISPLACEMENT",
+                    "value": [
+                        0.0,
+                        0.0,
+                        0.0
+                    ],
+                    "constrained": [
+                        true,
+                        true,
+                        true
+                    ],
+                    "interval": [
+                        0.0,
+                        "End"
+                    ]
+                }
+            }
+        ],
+        "loads_process_list": [],
+        "list_other_processes": []
+    },
+    "output_processes": {},
+    "analysis_stage": "KratosMultiphysics.SystemIdentificationApplication.sensor_sensitivity_solvers.sensor_sensitivity_static_analysis"
+}
