+++ applications/SystemIdentificationApplication/tests/responses/auxiliary_files/measured_data.csv
@@ -1,0 +1,2 @@
+#,type,name,location_0,location_1,location_2,value
+1,displacement_sensor,disp_x_1,0.75,0.75,0.0,0.20331919552457928
