+++ applications/SystemIdentificationApplication/tests/responses/auxiliary_files/primal_project_parameters.json
@@ -1,0 +1,109 @@
+{
+    "problem_data"     : {
+        "problem_name"  : "gg",
+        "parallel_type" : "OpenMP",
+        "echo_level"    : 1,
+        "start_time"    : 0.0,
+        "end_time"      : 1.0
+    },
+    "solver_settings"  : {
+        "time_stepping"                        : {
+            "time_step" : 1.1
+        },
+        "solver_type"                          : "Static",
+        "model_part_name"                      : "Structure",
+        "domain_size"                          : 2,
+        "echo_level"                           : 0,
+        "analysis_type"                        : "non_linear",
+        "model_import_settings"                : {
+            "input_type"     : "use_input_model_part"
+        },
+        "material_import_settings"             : {
+            "materials_filename" : "auxiliary_files/StructuralMaterials.json"
+        },
+        "line_search"                          : false,
+        "convergence_criterion"                : "residual_criterion",
+        "displacement_relative_tolerance"      : 0.0001,
+        "displacement_absolute_tolerance"      : 1e-9,
+        "residual_relative_tolerance"          : 0.0001,
+        "residual_absolute_tolerance"          : 1e-9,
+        "max_iteration"                        : 10,
+        "use_old_stiffness_in_first_iteration" : false,
+        "rotation_dofs"                        : true,
+        "volumetric_strain_dofs"               : false
+    },
+    "processes"        : {
+        "constraints_process_list" : [
+        {
+            "python_module": "assign_vector_variable_process",
+            "kratos_module": "KratosMultiphysics",
+            "help": "This process fixes the selected components of a given vector variable",
+            "process_name": "AssignVectorVariableProcess",
+            "Parameters": {
+                "mesh_id": 0,
+                "model_part_name": "Structure.DISPLACEMENT_left",
+                "variable_name": "DISPLACEMENT",
+                "value": [
+                    0.0,
+                    0.0,
+                    0.0
+                ],
+                "constrained": [
+                    true,
+                    true,
+                    true
+                ],
+                "interval": [
+                    0.0,
+                    "End"
+                ]
+            }
+        }],
+        "loads_process_list"       : [
+        {
+            "python_module": "assign_vector_by_direction_to_condition_process",
+            "kratos_module": "KratosMultiphysics",
+            "help": "This process sets a vector variable value over a condition",
+            "check": "DirectorVectorNonZero direction",
+            "process_name": "AssignModulusAndDirectionToConditionsProcess",
+            "Parameters": {
+                "model_part_name": "Structure.LineLoad2D_right",
+                "variable_name": "LINE_LOAD",
+                "modulus": 0.01,
+                "direction": [
+                    1.0,
+                    0.0,
+                    0.0
+                ],
+                "interval": [
+                    0.0,
+                    "End"
+                ]
+            }
+        }],
+        "list_other_processes"     : []
+    },
+    "output_processes" : {
+        "vtk_output" : [{
+            "python_module" : "vtk_output_process",
+            "kratos_module" : "KratosMultiphysics",
+            "process_name"  : "VtkOutputProcess",
+            "Parameters"    : {
+                "model_part_name"                             : "Structure",
+                "output_control_type"                         : "step",
+                "output_interval"                             : 1,
+                "file_format"                                 : "ascii",
+                "output_precision"                            : 7,
+                "output_sub_model_parts"                      : false,
+                "output_path"                                 : "vtk_output",
+                "save_output_files_in_folder"                 : true,
+                "nodal_solution_step_data_variables"          : ["DISPLACEMENT","REACTION"],
+                "nodal_data_value_variables"                  : [],
+                "element_data_value_variables"                : [],
+                "condition_data_value_variables"              : [],
+                "gauss_point_variables_extrapolated_to_nodes" : ["VON_MISES_STRESS"]
+            }
+        }]
+    },
+    "analysis_stage"   : "KratosMultiphysics.StructuralMechanicsApplication.structural_mechanics_analysis"
+}
